rename: web_view folder
