+++ lib/core/routing/router.dart
@@ -9,11 +9,11 @@ import 'package:dongsoop/presentation/home/notice_list_page_screen.dart';
 import 'package:dongsoop/presentation/main/main_screen.dart';
 import 'package:dongsoop/presentation/my_page/my_page_screen.dart';
 import 'package:dongsoop/presentation/schedule/schedule_screen.dart';
-import 'package:dongsoop/presentation/webview/cafeteria_web_view_page_screen.dart';
-import 'package:dongsoop/presentation/webview/library_banner_webview_screen.dart';
-import 'package:dongsoop/presentation/webview/notice_webview_screen.dart';
 import 'package:dongsoop/presentation/sign_in/sign_in_screen.dart';
 import 'package:dongsoop/presentation/sign_up/sign_up_screen.dart';
+import 'package:dongsoop/presentation/web_view/cafeteria_web_view_page_screen.dart';
+import 'package:dongsoop/presentation/web_view/library_banner_webview_screen.dart';
+import 'package:dongsoop/presentation/web_view/notice_webview_screen.dart';
 import 'package:go_router/go_router.dart';
 
 final router = GoRouter(initialLocation: RoutePaths.home, routes: [
@@ -82,7 +82,7 @@ final router = GoRouter(initialLocation: RoutePaths.home, routes: [
                   path: RoutePaths.libraryWebView,
                   name: 'libraryWebView',
                   builder: (context, state) =>
-                  const LibraryBannerWebViewScreen(),
+                      const LibraryBannerWebViewScreen(),
                 ),
                 GoRoute(
                   path: RoutePaths.cafeteriaWebView,
